Fix: added page wrongly treated as empty for some time.
When page is saved, file becomes non empty, but cached FastFile is not updated.
We don't know the real size of file until query, but we does not need actual file, just want to know whether it's empty or not.
So we assign dummy non-zero size if we save the file.