Merging background image feature, but not yet implement for CanvasBoox.
This merge is for picking up later changing.
We'll support it later in android.

Below is original commit log of chery pick.
----
Implement background image feature for ruled line.

If there is the png file named background.png, this app draw this png first, then draw next with blendingMode multiply.
This is for ruled-line template like feature.